refactor: split daemon service module
